// client/src/constants/colors.ts
// Make a color enum

export enum Color {
    WHITE = "#ffffff",
    LIGHTBLUE = "#add8e6",
    BABYPINK = "#ffb6c1",
    LAVENDER = "#e6e6fa",
    MINTGREEN = "#f5fffa",
    PALEGREEN = "#98fb98",
    LEMONCHIFFON = "#fffacd",
    PEACHPUFF = "#ffdab9",
    LIGHTCORAL = "#f08080",
    THISTLE = "#d8bfd8"
}